dpdp-pii-redaction-platform/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.2
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
loguru==0.7.2
python-dotenv==1.0.0
celery==5.3.6
redis==5.0.1
spacy==3.7.2
presidio-analyzer==2.2.353
presidio-anonymizer==2.2.353
